Add patch to remove RUNPATH from --with-advance-toolchain on GCC 11
GCC adds AT's RUNPATH to every binary it compiles when built with the
option '--with-advance-toolchain'. A GCC configured this way will fail
to build GlibC. The RUNPATH was required in older AT's but now became
obsolete for the ld.so.cache from AT with the correct DT_INTERP.

This patch is supposed to enter on GCC upstream on GCC 12.